implementation of devices for stm32f4 and stm32f1

+++ hardware/include/gpio.h
@@ -28,7 +28,7 @@ typedef enum {
 
 void gpio_config(pin_t pin, pin_dir_t dir);
 void gpio_set(pin_t pin, int enabled);
-int gpio_get(pin_t pin);
+uint8_t gpio_get(pin_t pin);
 
 inline uint8_t get_port(pin_t pin);
 inline uint8_t get_pin(pin_t pin);
