+++ hardware/src/gpio-lpc17xx.c
@@ -29,6 +29,6 @@ void gpio_set(pin_t pin, int enabled) {
         FIO_ClearValue(get_port(pin), 1 << get_pin(pin));
 }
 
-int gpio_get(pin_t pin) {
+uint8_t gpio_get(pin_t pin) {
     return (FIO_ReadValue(get_port(pin)) & (1 << get_pin(pin))) ? 1 : 0;
 }
